+++ README.md
@@ -27,3 +27,16 @@ and to fix all auto-fixable linting issues
 ```
 npm run lint -- --fix
 ```
+
+
+## Project Structure
+
+```
+```
+- src/
+  - main/     node code
+  - renderer/ vue code
+  - shared/   code shared between node and vue
+  - preload/  expose node code to vue code
+```
+  ```
