+++ index.html
@@ -6,6 +6,6 @@
   </head>
   <body>
     <div id="app"></div>
-    <script type="module" src="/src/renderer.js"></script>
+    <script type="module" src="/src/renderer/renderer.js"></script>
   </body>
 </html>
