--- src/index.css
@@ -1,17 +1,0 @@
-body {
-  font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, Helvetica,
-    Arial, sans-serif;
-  margin: 1em;
-  padding: 0;
-}
-
-.checkered {
-  background-image: repeating-conic-gradient(#606060 0% 25%, #bdbdbd 0% 50%);
-  background-position: 0 0, 45px 45px;
-  background-size: 90px 90px;
-  background-color: #bdbdbd;
-  width: max-content;
-  height: max-content;
-  overflow: hidden;
-  border: 1px solid #000000;
-}
